Refactor cli.py into focused modules: _cli_utils and _cli_sections
Split the 2938-line cli.py monolith into three files with clear
single responsibilities, preserving all behavior and tests exactly:

- gutenbit/_cli_utils.py (~249 lines): Pure utility functions with no
  side effects beyond stdout — string helpers, FTS query utilities,
  table-printing, formatting, and package version lookup.

- gutenbit/_cli_sections.py (~639 lines): The section/TOC data model —
  TypedDicts, title analysis, section path utilities, opening-row
  helpers, and the full _build_section_summary pipeline used by both
  the toc and view commands.

- gutenbit/cli.py (~2180 lines): The Click interface — all command
  definitions, CLI-specific helpers, JSON output functions, ingest
  helpers, and the entry point. Imports the extracted symbols so
  existing test imports (e.g. _select_section_opening_line,
  _build_section_summary) continue to resolve from gutenbit.cli.

No logic, signatures, or command interfaces were changed.
All 330 tests pass.
